chore(Chatbot): Integrate Chatbot Attachment into Chatbot demo
Renamed menu demo and merged the other two demos since they don't need to be separate - it will be easier for users to see features if they are in fewer places.
