+++ packages/module/patternfly-docs/content/extensions/virtual-assistant/examples/ChatbotAttachmentMenu/ChatbotAttachmentMenu.md
@@ -5,7 +5,7 @@ section: extensions
 subsection: Chat bots / AI
 # Sidenav secondary level section
 # should be the same for all markdown files
-id: Chatbot attachment with menu
+id: Chatbot demo with attachment menu
 # Tab (react | react-demos | html | html-demos | design-guidelines | accessibility)
 source: react
 # If you use typescript, the name of the interface to display props for
